test: share the diff-settling helper instead of copying it
`settleDiff` existed twice — `CommitPanel.lineFocus.test.tsx` had it first, and
de-flaking the line-staging suite added a second copy. It is subtle enough that
two copies drift, so it moves to `src/test/settle.ts` and both callers import it.

The doc comment travels with it, including the line-focus original's explanation
of WHY every new diff resets the cursor, plus the passive-effect mechanism the
line-staging fix uncovered. Stranding that in one caller was the reason the same
race had to be diagnosed twice.
